+++ database/migrations/2025_06_12_011407_create_auctions_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('auctions')) {
+            return;
+        }
+
         Schema::create('auctions', function (Blueprint $table) {
             $table->id();
             $table->foreignId('book_id')->constrained()->onDelete('cascade');
