+++ database/migrations/2025_06_20_001139_create_auction_users_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('auction_users')) {
+            return;
+        }
+
         Schema::create('auction_users', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
